runtime: Allow installing a different containerd
As done by the CoCo Operator, let's ensure that folks are allowed to
replace the containerd from the node using our helm chart.

Although you really should be careful doing this, there are situations
where it makes sense, as in ... you maintain your own infra and want a
patched / newer version of containerd than what your OS provides you.

This change is entirely based on what I did for the Operator in the
past, and vibe generated using Cursor AI.

Signed-off-by: Fabiano Fidêncio <ffidencio@nvidia.com>