tests: e2e: Retry pod creation to rule out slow network
Slow network can cause the pod to never come up as there's a 2 minutes
limit on the kuberentes side.

As the test is passing the majority of the cases, let's add a retry
mechanism so we don't fail on the first failure, avoiding to have to
re-run the whole test again.

Signed-off-by: Fabiano Fidêncio <ffidencio@nvidia.com>