charts: Use `coco-system` as the default namespace
It's bad practice to install anything in the kube-system namespace,
unless it actually needs to run there - and users may install coco in
kube-system simply because the documentation defaults to kube-system.

This makes sense from a security point of view, and also from a
maintenance point of view, should CoCo blow up it's at least confined to
it's own namespace making containment and cleanup a lot easier.

Fixes: #24

Signed-off-by: Fabiano Fidêncio <ffidencio@nvidia.com>
Suggested-by: Lasse Østerild <lasse@oesterild.dk>